requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
Pillow==10.0.1
PyJWT==2.8.0
PyPDF2==1.26.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.6
Werkzeug==3.0.1
yarl==1.9.2
